# docs/swagger.yaml
definitions:
...
  dto.CreateAccessPermissionPayloadDTO:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      parent_id:
        type: string
    required:
    - code
    - name
    type: object
  dto.CreateAccessPolicyPayloadDTO:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      parent_id:
        type: string
    required:
    - code
    - name
    type: object
...
  dto.PaginationDTOResponse:
    properties:
      data:
        items: {}
        type: array
      next_page:
        type: integer
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  dto.PermissionDTO:
    properties:
      action:
        description: The action this permission allows
        type: string
      created_at:
        description: Timestamp of permission creation
        type: string
      description:
        description: Optional description of the permission
        type: string
      id:
        description: Unique ID for the permission
        type: string
      policy_id:
        description: Foreign key referencing IAMPolicy
        type: string
      resource:
        description: The resource this permission applies to
        type: string
      updated_at:
        description: Timestamp of last update
        type: string
    type: object
...
  dto.PolicyDTO:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.UpdateAccessPermissionPayloadDTO:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      parent_id:
        type: string
    required:
    - code
    - name
    type: object
  dto.UpdateAccessPolicyPayloadDTO:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      parent_id:
        type: string
    required:
    - code
    - name
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: string
      errors:
        items: {}
        type: array
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/permissions:
    get:
      consumes:
      - application/json
      description: Get permissions
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      - description: Keyword
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful retrieval of permissions
          schema:
            $ref: '#/definitions/dto.PaginationDTOResponse'
        "400":
          description: Invalid page number or size
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Retrieve permissions
      tags:
      - permissions
    post:
      consumes:
      - application/json
      description: Create a new permission
      parameters:
      - description: permission payload
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAccessPermissionPayloadDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Successful creation of permission
          schema:
            $ref: '#/definitions/dto.PermissionDTO'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a new permission
      tags:
      - permissions
  /api/v1/permissions/{permission_id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing permission
      parameters:
      - description: permission ID
        in: path
        name: permission_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successful deletion of permission
        "400":
          description: Invalid request ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: permission not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete an existing permission
      tags:
      - permissions
    get:
      consumes:
      - application/json
      description: Get permission by ID
      parameters:
      - description: permission ID
        in: path
        name: permission_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful retrieval of permission
          schema:
            $ref: '#/definitions/dto.PermissionDTO'
        "400":
          description: Invalid request ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: permission not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Retrieve permission by ID
      tags:
      - permissions
    put:
      consumes:
      - application/json
      description: Update an existing permission
      parameters:
      - description: permission ID
        in: path
        name: permission_id
        required: true
        type: string
      - description: permission payload
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAccessPermissionPayloadDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successful update of permission
          schema:
            $ref: '#/definitions/dto.PermissionDTO'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: permission not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update an existing permission
      tags:
      - permissions
  /api/v1/policies:
    get:
      consumes:
      - application/json
      description: Get policies
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      - description: Keyword
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful retrieval of policies
          schema:
            $ref: '#/definitions/dto.PaginationDTOResponse'
        "400":
          description: Invalid page number or size
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Retrieve policies
      tags:
      - policies
    post:
      consumes:
      - application/json
      description: Create a new policy
      parameters:
      - description: policy payload
        in: body
        name: policy
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAccessPolicyPayloadDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Successful creation of policy
          schema:
            $ref: '#/definitions/dto.PolicyDTO'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a new policy
      tags:
      - policies
  /api/v1/policies/{policy_id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing policy
      parameters:
      - description: policy ID
        in: path
        name: policy_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successful deletion of policy
        "400":
          description: Invalid request ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: policy not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete an existing policy
      tags:
      - policies
    get:
      consumes:
      - application/json
      description: Get policy by ID
      parameters:
      - description: policy ID
        in: path
        name: policy_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful retrieval of policy
          schema:
            $ref: '#/definitions/dto.PolicyDTO'
        "400":
          description: Invalid request ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: policy not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Retrieve policy by ID
      tags:
      - policies
    put:
      consumes:
      - application/json
      description: Update an existing policy
      parameters:
      - description: policy ID
        in: path
        name: policy_id
        required: true
        type: string
      - description: policy payload
        in: body
        name: policy
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAccessPolicyPayloadDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successful update of policy
          schema:
            $ref: '#/definitions/dto.PolicyDTO'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: policy not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update an existing policy
      tags:
      - policies
...
swagger: "2.0"